perf(nestjs-util): rewrite filters
BREAKING CHANGE: - GlobalExceptionFilter stays in the @webundsoehne/nestjs-util
- BadRequestExceptionFilter is deleted since it now utilizes ClassValidatorException and
ExtendedValidationPipe to throw the error directly as HttpException
- RpcGlobalExceptionFilter is deleted and GlobalExceptionFilter can be used for it directly since
with the newer nestjs versions it seems to parse the context correctly
- From anywhere in your code you can throw ClassValidatorException to throw arbitrary validation
errors if you do validateOrReject().
- error stack trace is propagated to the service directly (see 1th screenshot) and it will write the
service name in front if it to designate it as external error (for now only configrued for rmq)
- seems to handle all kinds of messages i can think of see rest of the screenshots
- refactored the class validator error a bit more. for child properties it includes the parent
property name like parent.child.test etc. (see last screenshots)