fix(nx-executors): update build flow
Use the "default" build flow with a `dist` folder in the package. Include `plugin` folder with js files directly without the need of building that folder first. This supports our build & release flow (no need to switch pkgRoot) and allows to access the plugins without the `/dist` in the path.